base/rag/indexers/architecture/requirements.txt
pymilvus>=2.5.0
httpx>=0.28.0
pyyaml>=6.0
pydantic>=2.0
PyMuPDF>=1.24.0
beautifulsoup4>=4.12.0
markdownify>=0.13.0
